# shared/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hydroponic Agri Management Backend API
  version: 0.1.0
  description: Browser API documentation for hydroponic backend services.
servers:
  - url: http://127.0.0.1:8080
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/alerts/subscribe:
    get:
      tags: [alerts]
      summary: Alert subscribe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
...